[bar-jhamf6] Доработал функцию обновления токена.
