+++ src/auction/auction.js
@@ -6,10 +6,12 @@ import $ from "jquery";
 import VueRouter from 'vue-router';
 import axios from 'axios';
 
-$(function () {    
-    // TODO: Переделать на другой способ глобального хранения! 
-    __VUE_HOT_MAP__.refreshToken();
-});
+import { refreshToken } from '../store.js';
+
+// $(function () {    
+//     // TODO: Переделать на другой способ глобального хранения! 
+//     __VUE_HOT_MAP__.refreshToken();
+// });
 
 export default {
     name: 'auction',
@@ -19,6 +21,7 @@ export default {
     created() {
         this._loadingAuctionTasks();
         this._totalPageetPagination();
+        refreshToken();
     },
     data() {
         return {
