+++ src/customer-header/customer-header.js
@@ -6,10 +6,12 @@ import $ from "jquery";
 import axios from 'axios';
 import VueRouter from 'vue-router';
 
-$(function () {
-    // TODO: Переделать на другой способ глобального хранения! 
-    __VUE_HOT_MAP__.refreshToken();
-});
+import { refreshToken } from '../store.js';
+
+// $(function () {
+//     // TODO: Переделать на другой способ глобального хранения! 
+//     __VUE_HOT_MAP__.refreshToken();
+// });
 
 export default {
     name: 'customer-header',
@@ -28,6 +30,7 @@ export default {
     created() {
         this.bGuest = sessionStorage["role"] == "G" ? true : false;
         this._getBalance();
+        refreshToken();
     },
     methods: {
         // Функция проверяет авторизован ли юзер.
