+++ src/my-tasks/my-tasks.js
@@ -8,10 +8,12 @@ import $ from "jquery";
 import VueRouter from 'vue-router';
 import axios from 'axios';
 
-$(function () {    
-    // TODO: Переделать на другой способ глобального хранения! 
-    __VUE_HOT_MAP__.refreshToken();
-});
+import { refreshToken } from '../store.js';
+
+// $(function () {    
+//     // TODO: Переделать на другой способ глобального хранения! 
+//     __VUE_HOT_MAP__.refreshToken();
+// });
 
 export default {
     name: 'my-tasks',
@@ -25,6 +27,7 @@ export default {
         this._totalPageetPagination();
         this.onGetPagination(1);
         this._loadMyTasks();
+        refreshToken();
     },
     data() {
         return {
