+++ src/nav-menu/nav-menu.js
@@ -7,10 +7,12 @@ import $ from "jquery";
 import VueRouter from 'vue-router';
 import axios from 'axios';
 
-$(function () {    
-    // TODO: Переделать на другой способ глобального хранения! 
-    __VUE_HOT_MAP__.refreshToken();
-});
+import { refreshToken } from '../store.js';
+
+// $(function () {    
+//     // TODO: Переделать на другой способ глобального хранения! 
+//     __VUE_HOT_MAP__.refreshToken();
+// });
 
 export default {
     name: 'nav-menu',
@@ -20,7 +22,8 @@ export default {
     },
     created() {
         this._getAuthorize();   
-        this._initHeader();                
+        this._initHeader();      
+        refreshToken();          
     },
     data() {
         return {
