+++ src/view-task/view-task.js
@@ -9,10 +9,12 @@ import axios from 'axios';
 import Calendar from 'v-calendar/lib/components/calendar.umd';
 import DatePicker from 'v-calendar/lib/components/date-picker.umd';
 
-$(function () {    
-    // TODO: Переделать на другой способ глобального хранения! 
-    __VUE_HOT_MAP__.refreshToken();
-});
+import { refreshToken } from '../store.js';
+
+// $(function () {    
+//     // TODO: Переделать на другой способ глобального хранения! 
+//     __VUE_HOT_MAP__.refreshToken();
+// });
 
 export default {
     name: 'view-task',
@@ -25,7 +27,8 @@ export default {
 
     created() {
         this._loadingResponds();
-        this._loadingDialogs();        
+        this._loadingDialogs();   
+        refreshToken();     
     },
 
     data() {
